rimage: Allow blocks to be loaded into specific memory areas
This will allow firmware binaries to have sections that are
loaded in different bars.

Memory areas are "manually" specified inside struct adsp. We
change that with an array of memory zones indexed by BAR type.

The indexes are specified via snd_sof_fw_blk_type which now
encodes the BAR type. To keep backward compatibility we map:
	-> SOF_BLK_TEXT to SOF_BAR_TYPE_IRAM
	-> SOF_BLK_DATA to SOF_BAR_TYPE_DRAM.

We get rid of SOF_BLK_{CACHE|REGS|SIGS} because they are not
used in SOF firmware (and ignored on Linux).

A minor drawback of this patch is that each machine description
will have memory allocated for all existing BARs even if less
are used. This can be refined in the future, by dynamically
allocating the memory zones.

Suggested-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>
Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>