WIP: move to shimmed codec adapter interface
This commit strips the codec adapter as it is more of a library/layer
for codecs to simplify their interface against rather than a component
in and of itself. This gives each codec its own UUID solving the ID
problem while opening an opertunity to register codecs directly for
future multiplexing capabilities.

Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>