Audio: Remove DMA operations from simulated Zephyr DAI
With this change the DAI just consumes all playback samples
and produces triangle wave capture samples.

In the playback direction the consumed samples are checked for
glitches base on assumed triangle wave. There is also count of
leading zero PCM samples per channel. It can be used to estimate
latency in the playback path.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>